requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.44
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.11
markupsafe==2.1.3
